+++ bobsled/storages/__init__.py
@@ -1,0 +1,2 @@
+from .database import DatabaseStorage  # noqa
+from .memory import InMemoryStorage  # noqa
